requirements.txt
langgraph>=0.2.0
psycopg-pool>=3.0.0
python-dotenv>=1.0.0
langchain-openai>=0.1.0
langchain-community>=0.0.1
langchain-tavily>=0.1.0
pydantic>=2.0.0
loguru>=0.7.0
langsmith>=0.1.0
mcp>=1.0.0
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0
# 嵌入模型和向量存储
langchain-huggingface>=0.1.0
transformers>=4.38.0
sentence-transformers>=2.2.0
huggingface-hub>=0.33.4,<1.0.0
# pgvector 支持
pgvector>=0.2.0
# PyTorch with CUDA 12.4 support
# Install with: pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu124
torch>=2.6.0
torchvision>=0.21.0
torchaudio>=2.6.0

# 小红书 MCP 依赖
xiaohongshu-automation>=2.0.0
pandas>=2.0.0
playwright>=1.47.0
fastmcp>=2.14.0

# Async dependencies for TravelAgent async refactoring
asyncpg>=0.29.0
aiosqlite>=0.19.0
httpx>=0.25.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic-settings>=2.0.0
